finished debugging mobile controls on android
