+++ COMP7051_Asg02/Assets/_Scripts/PlayerController.cs
@@ -32,27 +32,43 @@ public class PlayerController : MonoBehaviour {
 
         rb = GetComponent<Rigidbody>();
     }
-
-    // Update is called once per frame
+    /*
     void Update() {
+        if (Input.touchCount == 1) {
+            //forward movement
+            Touch touchZero = Input.GetTouch(0);
+            float moveSpeed = touchZero.position.y / Screen.height;
+            transform.Translate(transform.forward * moveSpeed * walkSpeed * Time.deltaTime, Space.World);
+
+
+            //rotation
+            float rotateSpeed = touchZero.position.x / Screen.width - .5f;
+            transform.Rotate(0, rotateSpeed * turnSpeed * Time.deltaTime, 0);
+            anim.SetFloat("forwardMotion", 1);
+        }
+    }
+    */
+        // Update is called once per frame
+        void Update() {
         if (Input.GetAxis("Vertical") > .1)
             anim.SetFloat("forwardMotion", Input.GetAxis("Vertical"));
         else
             anim.SetFloat("forwardMotion", 0);
 
         //if run button is held down enable running
-        if (IRefs.GetKey(IRefs.Command.PlayerRun))
+        if (Application.platform != RuntimePlatform.Android && IRefs.GetKey(IRefs.Command.PlayerRun))
             isRunning = true;
         else
             isRunning = false;
 
         anim.SetBool("run", isRunning);
+
         float speed = isRunning ? runSpeed : walkSpeed;
 
         //forward movement (controller and keyboard)
         if (Input.GetAxis("Vertical") > 0)
             transform.Translate(transform.forward * Input.GetAxis("Vertical") * speed * Time.deltaTime, Space.World);
-        
+
         //rotation (controller and keyboard)
         if (Input.GetAxis("Horizontal") != 0) {
             transform.Rotate(0, Input.GetAxis("Horizontal") * turnSpeed * Time.deltaTime, 0);
@@ -60,7 +76,7 @@ public class PlayerController : MonoBehaviour {
         }
 
         //mouse controls for character movement and rotation
-        if (Input.GetMouseButton(0) && Input.touchCount == 0) {
+        if (Input.touchCount == 0 && Input.GetMouseButton(0)) {
             transform.Translate(transform.forward * speed * Time.deltaTime, Space.World);
 
             float rotationDirection = Input.mousePosition.x / (Screen.width / 2) - 1;
@@ -71,12 +87,13 @@ public class PlayerController : MonoBehaviour {
         //touchscreen controls movement and rotation
         if (Input.touchCount == 1) {
             //forward movement
-            float moveSpeed = Input.GetTouch(0).position.y / Screen.height;
+            Touch touchZero = Input.GetTouch(0);
+            float moveSpeed = touchZero.position.y / Screen.height;
             transform.Translate(transform.forward * moveSpeed * speed * Time.deltaTime, Space.World);
 
 
             //rotation
-            float rotateSpeed = Input.GetTouch(0).position.x / Screen.width - .5f;
+            float rotateSpeed = touchZero.position.x / Screen.width - .5f;
             transform.Rotate(0, rotateSpeed * turnSpeed * Time.deltaTime, 0);
             anim.SetFloat("forwardMotion", 1);
         }
@@ -85,7 +102,7 @@ public class PlayerController : MonoBehaviour {
             toggleDetectCollisions();
 
     }
-
+    
     /**toggles ability to walk through walls*/
     public void toggleDetectCollisions() {
         detectCollisions = !detectCollisions;
